backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.9
redis==5.0.7
APScheduler==3.10.4
cryptography==43.0.1
slowapi==0.1.9
orjson==3.10.7
httpx==0.27.2
tenacity==9.0.0
pytest==8.3.2
pytest-asyncio==0.23.8
